refactor(ui): modularise widgets
- Move `notifications`, `toolbar`, and `status_bar` to a new `widgets` module
- Replace `dock_state` with `editor_state` for clarity in `UiState`
- Enhance `render_editor_layout` to include status bar rendering
- Refactor `toolbar` and `status_bar` rendering functions
- Add a dedicated `debug_span` for `render_editor_layout`
- Expand logging filter configuration in `LogConfiguration::default`
- Remove deprecated localisation files and native-dialog dependency
