+++ crates/ui/src/state.rs
@@ -11,8 +11,9 @@ use egui_dock::{DockState, NodeIndex};
 #[derive(Resource)]
 pub struct UiState {
     /// The [`DockState`](https://docs.rs/egui_dock/latest/egui_dock/dock_state/struct.DockState.html)
-    /// that controls most of the general layout.
-    pub dock_state: DockState<EditorPanels>,
+    ///
+    /// that controls most of the editing layout.
+    pub editor_state: DockState<EditorPanels>,
 }
 
 impl Default for UiState {
@@ -27,6 +28,8 @@ impl Default for UiState {
         );
         let [_, _settings] = surface.split_below(layers, 0.6, vec![EditorPanels::Settings]);
 
-        Self { dock_state: state }
+        Self {
+            editor_state: state,
+        }
     }
 }
