+++ crates/ui/src/widgets.rs
@@ -1,0 +1,5 @@
+//! This module contains various reusable widgets that can be shared across various UI layouts.
+
+pub mod notifications;
+pub mod status_bar;
+pub mod toolbar;
