--- locales/en-GB/assets.ftl
@@ -1,6 +1,0 @@
-# Format: {domain}-{component}-{element}-{variant}
-
-assets-add_pack-packs-header = Asset Packs
-assets-add_pack-add_button = add new pack
-assets-add_pack-added-notification = Asset pack {name} added
-assets-add_pack-adding_failed-notification = Failed to add asset pack: {error}
