--- locales/en-GB/dialogs.ftl
@@ -1,3 +1,0 @@
-# Format: {domain}-{component}-{element}-{variant}
-
-dialogs-new-project-title = Create new project
